backend/requirements.txt
# djanggo library
Django>=4.2,<5.0
djangorestframework
django-cors-headers
django-environ

# database library
psycopg2-binary
pgvector

# production server
gunicorn

# Utilities
python-dotenv
requests
httpx

python-decouple sendgrid
